Break down CmdMigrationMonitor into separate classes
